agent-patterns/pyproject.toml
[project]
name = "agent-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.13",
]
